debezium/dbz#2288 Consolidate tests around shared pipeline infrastructure
Test-only restructuring, no production code changes:

- Add AbstractCockroachDBPipelineIT owning the Testcontainers stack and the
  in-process connector task plumbing that pipeline ITs previously copied.
- Fold the five incident-named regression ITs (decimal precision, delete key,
  NOT NULL JSONB backlog, changefeed diff reuse, kafka sink config) into
  CockroachDBRegressionScenariosIT, which runs them against one shared stack
  with per-scenario databases and topic prefixes.
- Fold the incident-named unit classes into the component classes that own the
  code under test: decimal precision and delete key tests into
  CockroachDBChangeRecordEmitterTest, NOT NULL column conversion tests into
  CockroachDBValueConverterProviderTest, and the dedup and kafka sink config
  tests into a new CockroachDBStreamingChangeEventSourceTest.

All 26 moved test methods verified present in their new homes; 248 unit tests
and 47 integration tests pass, the same counts as before the restructuring.
